(refactor) db creds/functions  moved to single files respectively
